ScraperService/requirements.txt
selenium==4.11.2
python-dateutil==2.8.1
bs4==0.0.1
beautifulsoup4==4.12.3
certifi==2023.7.22
pandas==1.2.1
Django==4.2.13
Faker==26.0.0
djangorestframework==3.15.2
lxml==4.6.2
soupsieve==2.1
dnspython==2.3.0
trio==0.22.0
idna==2.10
requests==2.25.1
h11==0.14.0
pip==24.1.2
configparser==5.0.1
attrs==23.1.0
wheel==0.40.0
Jinja2==3.1.2
numpy==1.26.4
blis==0.7.9
pytz==2021.1
tqdm==4.59.0
colorama==0.4.4
outcome==1.2.0
sniffio==1.3.0
exceptiongroup==1.1.1
sortedcontainers==2.4.0
PyYAML==6.0.1
cymem==2.0.7
setuptools==67.6.1
pathy==0.10.1
typer==0.7.0
regex==2023.8.8
thinc==8.1.12
srsly==2.4.6
wasabi==1.1.2
click==8.1.3
catalogue==2.0.8
pydantic==1.10.9
packaging==23.1
torch==2.0.1
langcodes==3.3.0
preshed==3.0.8
sympy==1.12
mpmath==1.3.0
murmurhash==1.0.9
confection==0.0.4
transformers==4.30.2
urllib3==1.26.3
asgiref==3.8.1
sqlparse==0.5.0
psycopg2==2.8.4
fsspec==2023.6.0
MarkupSafe==2.1.3
joblib==1.3.2
appdirs==1.4.4
six==1.15.0
pipenv==2020.11.15
distlib==0.3.1
chardet==4.0.0
zipp==3.4.0
python-dotenv==0.15.0
crayons==0.4.0
gunicorn==20.0.4
networkx==3.1
tokenizers==0.13.3
virtualenv==20.4.2
filelock==3.0.12
safetensors==0.3.3
wsproto==1.2.0
## The following requirements were added by pip freeze:
async-generator==1.10
charset-normalizer==2.0.4
chromedriver-autoinstaller==0.6.2
cloudpathlib==0.18.1
dj-config-url==0.1.1
dj-database-url==0.4.1
django-cors-headers==3.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.5.0/es_core_news_sm-3.5.0-py3-none-any.whl#sha256=13b909bac0de5f5338f4327b14a9cc9736cf51f07131a3cfcb688dbeb8608bdd
es-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/es_dep_news_trf-3.5.0/es_dep_news_trf-3.5.0-py3-none-any.whl#sha256=4322e809c35ae135cc8fa71831c1aeafc1b6232886c45b247ed431d264d3445e
export==0.1.2
huggingface-hub==0.16.4
importlib-metadata==3.4.0
pillow==10.4.0
psycopg2-binary==2.8.4
PySocks==1.7.1
python-Levenshtein==0.12.2
routers==0.10.1
smart-open==6.3.0
trio-websocket==0.10.2
typing_extensions==4.6.3
virtualenv-clone==0.5.4
weasel==0.4.1
webdriver-manager==4.0.1
whitenoise==5.2.0
